+++ src/nl/hannahsten/texifyidea/structure/latex/LatexLabelPresentation.kt
@@ -1,25 +1,15 @@
 package nl.hannahsten.texifyidea.structure.latex
 
 import com.intellij.navigation.ItemPresentation
-import com.intellij.openapi.fileEditor.FileDocumentManager
 import nl.hannahsten.texifyidea.TexifyIcons
-import nl.hannahsten.texifyidea.psi.LatexCommands
 
 /**
  * @author Hannah Schellekens
  */
-class LatexLabelPresentation(labelCommand: LatexCommands, labelPosition: Int) : ItemPresentation {
-
-    private val locationString: String
-    private val presentableText: String = labelCommand.requiredParameterText(labelPosition) ?: "no label found"
-
-    init {
-        // Location string.
-        val manager = FileDocumentManager.getInstance()
-        val document = manager.getDocument(labelCommand.containingFile.virtualFile)
-        val line = document!!.getLineNumber(labelCommand.textOffset) + 1
-        this.locationString = labelCommand.containingFile.name + ":" + line
-    }
+class LatexLabelPresentation(
+    private val locationString: String,
+    private val presentableText: String
+) : ItemPresentation {
 
     override fun getPresentableText() = presentableText