+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 * Improve many inspections, by @Ezrnest
+* Improve support for custom label commands, by @Ezrnest
 * Fold \left...\right expressions, by @jojo2357
 
 ### Fixed